general adjustments and exception handling

+++ src/main/java/com/guistraliote/attribute/AttributeService.java
@@ -1,10 +1,14 @@
 package com.guistraliote.attribute;
 
 
+import com.guistraliote.attribute.exceptions.AttributeNotFoundException;
+import com.guistraliote.attribute.exceptions.InvalidAttributeNameException;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.ws.rs.NotFoundException;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @ApplicationScoped
@@ -13,8 +17,10 @@ public class AttributeService {
     @Inject
     AttributeRepository attributeRepository;
 
-    public List<AttributeDTO> findAll() {
-        return attributeRepository.listAll().stream()
+    public List<AttributeDTO> findAllPaged(int pageIndex, int pageSize) {
+        List<Attribute> attributes = attributeRepository.findPaged(pageIndex, pageSize);
+
+        return attributes.stream()
                 .map(attribute -> AttributeDTO.builder()
                         .id(attribute.getId())
                         .attributeName(attribute.getAttributeName())
@@ -26,6 +32,11 @@ public class AttributeService {
 
     public AttributeDTO findById(Long id) {
         Attribute attribute = attributeRepository.findById(id);
+
+        if (Objects.isNull(attribute)) {
+            throw new AttributeNotFoundException("Attribute not found for ID: " + id);
+        }
+
         return AttributeDTO.builder()
                 .id(attribute.getId())
                 .attributeName(attribute.getAttributeName())
@@ -35,29 +46,50 @@ public class AttributeService {
     }
 
     public AttributeDTO create(AttributeDTO attributeDTO) {
+        if (checkAttributeName(attributeDTO)) {
+            throw new InvalidAttributeNameException("Attribute name cannot be null or empty.");
+        }
+
         Attribute attribute = Attribute.builder()
                 .attributeName(attributeDTO.getAttributeName())
                 .attributeValue(attributeDTO.getAttributeValue())
                 .isActive(attributeDTO.getIsActive())
                 .build();
+
         attributeRepository.persist(attribute);
-        attributeDTO.setId(attribute.getId());
+
         return attributeDTO;
     }
 
     public AttributeDTO update(Long id, AttributeDTO attributeDTO) {
-        Attribute attribute = attributeRepository.findById(id);
-        if (attribute != null) {
-            attribute.setAttributeName(attributeDTO.getAttributeName());
-            attribute.setAttributeValue(attributeDTO.getAttributeValue());
-            attribute.setIsActive(attributeDTO.getIsActive());
-            attributeRepository.persist(attribute);
-            return attributeDTO;
+        if (Objects.isNull(id)) {
+            throw new AttributeNotFoundException("Attribute ID cannot be null");
         }
-        return null;
+
+        Attribute attribute = attributeRepository.findById(id);
+
+        if (Objects.isNull(attribute)) {
+            throw new AttributeNotFoundException("Attribute not found for ID: " + id);
+        }
+
+        attribute.setAttributeName(attributeDTO.getAttributeName());
+        attribute.setAttributeValue(attributeDTO.getAttributeValue());
+        attribute.setIsActive(attributeDTO.getIsActive());
+
+        attributeRepository.persist(attribute);
+
+        return attributeDTO;
     }
 
     public void delete(Long id) {
-        attributeRepository.deleteById(id);
+        if (attributeRepository.findByIdOptional(id).isPresent()) {
+            attributeRepository.deleteById(id);
+        } else {
+            throw new AttributeNotFoundException("Attribute not found for ID: " + id);
+        }
+    }
+
+    private Boolean checkAttributeName(AttributeDTO dto) {
+        return Objects.isNull(dto.getAttributeName()) || dto.getAttributeName().trim().isEmpty();
     }
 }
